// src/types/oddsmatcher.ts
export interface OddsmatcherRow {
  sport: string
  home: string
  away: string
  date: string
  hour: string
  nation: string
  competition: string
  last_update: string
  id_book_1: string
  id_book_2: string
  url: string
  market: string
  back_odd: string
  lay_odd: string
  liquidity: string
  selection: string
}